requirements.txt
cloudscraper
joblib
optuna
pandas
python-dotenv
scikit-learn
streamlit
torch -f https://download.pytorch.org/whl/cu121